+++ packages/eslint-plugin/CHANGELOG.md
@@ -15,7 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- feat: Add `boundaries/flag-as-external` setting to allow better control over external module identification
+- feat(#349): Add `boundaries/flag-as-external` setting to allow better control over external module identification
+
+### Removed
+
+- feat(#423): Remove rules validation during linting runtime to improve performance. Validation is performed at configuration load time using eslint schema validation.
 
 ### Changed
 
@@ -25,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
-- fix(#415): Fix elements selector validation error when selector is an empty array
+- fix(#420): Fix elements selector validation error when selector is an empty array
 
 ## [5.3.0] - 2025-11-23